requirements.txt
# Core framework
streamlit>=1.40.0

# Excel processing
openpyxl>=3.1.2
pandas>=2.1.0

# LLM
anthropic>=0.40.0

# Data validation
pydantic>=2.5.0

# Fuzzy matching
rapidfuzz>=3.5.0

# Charts
plotly>=5.18.0

# Word export
python-docx>=1.1.0

# Environment variables
python-dotenv>=1.0.0
